# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Prompt Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  description: string;
  promptCount: number;
}

export interface Prompt {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  description?: string | null;
  categoryId: number;
  categoryName: string;
  tags: string[];
  authorUsername: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  saveCount: number;
  viewCount: number;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PromptListResponse {
  prompts: Prompt[];
  total: number;
}

export interface PromptInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  description?: string;
  categoryId: number;
  tags?: string[];
  authorUsername: string;
  isPublic?: boolean;
}

export interface PromptUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  content?: string;
  description?: string;
  categoryId?: number;
  tags?: string[];
  isPublic?: boolean;
}

export interface SaveToggleInput {
  username: string;
}

export interface SaveToggleResponse {
  saved: boolean;
  saveCount: number;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  categories: string[];
  promptCount: number;
  libraryCount: number;
  totalSaves: number;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 2 */
  username: string;
  /** @minLength 1 */
  displayName: string;
  bio?: string;
  avatarUrl?: string;
  categories?: string[];
}

export interface UserUpdate {
  /** @minLength 1 */
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  categories?: string[];
}

export interface Library {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  authorUsername: string;
  authorDisplayName: string;
  promptCount: number;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface LibraryDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  authorUsername: string;
  authorDisplayName: string;
  isPublic: boolean;
  prompts: Prompt[];
  createdAt: string;
  updatedAt: string;
}

export interface LibraryInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  authorUsername: string;
  isPublic?: boolean;
}

export interface LibraryUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  isPublic?: boolean;
}

export interface LibraryPromptInput {
  promptId: number;
}

export interface MarketplaceStats {
  totalPrompts: number;
  totalCreators: number;
  totalSaves: number;
  totalLibraries: number;
  topCategories: Category[];
}

export type ListPromptsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
sort?: ListPromptsSort;
/**
 * @nullable
 */
username?: string | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
offset?: number | null;
};

export type ListPromptsSort = typeof ListPromptsSort[keyof typeof ListPromptsSort] | null;


export const ListPromptsSort = {
  trending: 'trending',
  newest: 'newest',
  most_saved: 'most_saved',
} as const;

export type GetTrendingPromptsParams = {
/**
 * @nullable
 */
limit?: number | null;
};

export type GetFeaturedCreatorsParams = {
/**
 * @nullable
 */
limit?: number | null;
};

export type ListLibrariesParams = {
/**
 * @nullable
 */
username?: string | null;
/**
 * @nullable
 */
limit?: number | null;
};