# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prompt Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}/libraries:
    get:
      operationId: getUserLibraries
      tags: [users]
      summary: Get libraries for a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User libraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Library"
  /libraries:
    get:
      operationId: listLibraries
      tags: [libraries]
      summary: List public libraries
      parameters:
        - name: username
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of libraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Library"
    post:
      operationId: createLibrary
      tags: [libraries]
      summary: Create a new library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryInput"
      responses:
        "201":
          description: Created library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
  /libraries/{id}:
    get:
      operationId: getLibrary
      tags: [libraries]
      summary: Get a library by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Library with prompts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryDetail"
        "404":
          description: Not found
    patch:
      operationId: updateLibrary
      tags: [libraries]
      summary: Update a library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryUpdate"
      responses:
        "200":
          description: Updated library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
    delete:
      operationId: deleteLibrary
      tags: [libraries]
      summary: Delete a library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /libraries/{id}/prompts:
    post:
      operationId: addPromptToLibrary
      tags: [libraries]
      summary: Add a prompt to a library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryPromptInput"
      responses:
        "200":
          description: Updated library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
  /libraries/{id}/prompts/{promptId}:
    delete:
      operationId: removePromptFromLibrary
      tags: [libraries]
      summary: Remove a prompt from a library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: promptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
...
components:
  schemas:
...
    Prompt:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        tags:
          type: array
          items:
            type: string
        authorUsername:
          type: string
        authorDisplayName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        saveCount:
          type: integer
        viewCount:
          type: integer
        isPublic:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - content
        - categoryId
        - categoryName
        - tags
        - authorUsername
        - authorDisplayName
        - saveCount
        - viewCount
        - isPublic
        - createdAt
        - updatedAt
...
    Library:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        authorUsername:
          type: string
        authorDisplayName:
          type: string
        promptCount:
          type: integer
        isPublic:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - authorUsername
        - authorDisplayName
        - promptCount
        - isPublic
        - createdAt
        - updatedAt
    LibraryDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        authorUsername:
          type: string
        authorDisplayName:
          type: string
        isPublic:
          type: boolean
        prompts:
          type: array
          items:
            $ref: "#/components/schemas/Prompt"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - authorUsername
        - authorDisplayName
        - isPublic
        - prompts
        - createdAt
        - updatedAt
    LibraryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        authorUsername:
          type: string
        isPublic:
          type: boolean
      required:
        - name
        - authorUsername
    LibraryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        isPublic:
          type: boolean
    LibraryPromptInput:
      type: object
      properties:
        promptId:
          type: integer
      required:
        - promptId
...