# lib/api-zod/src/generated/types/library.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Prompt Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Library {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  authorUsername: string;
  authorDisplayName: string;
  promptCount: number;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}