# lib/api-zod/src/generated/types/libraryDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Prompt Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Prompt } from './prompt';

export interface LibraryDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  authorUsername: string;
  authorDisplayName: string;
  isPublic: boolean;
  prompts: Prompt[];
  createdAt: string;
  updatedAt: string;
}